minor fixes and using ranges

+++ thrust/examples/scan_matrix_by_rows.cu
@@ -14,7 +14,7 @@ __host__ void scan_matrix_by_rows0(thrust::device_vector<int>& u, int n, int m)
   // Here, we launch a separate scan for each row in the matrix. This works,
   // but each kernel only does a small amount of work. It would be better if we
   // could launch one big kernel for the entire matrix.
-  for (size_t i = 0; i < static_cast<size_t>(n); ++i)
+  for (int i = 0; i < n; ++i)
   {
     thrust::inclusive_scan(u.begin() + m * i, u.begin() + m * (i + 1), u.begin() + m * i);
   }
@@ -69,9 +69,9 @@ int main()
   thrust::sequence(u1.begin(), u1.end());
   scan_matrix_by_rows1(u1, n, m);
 
-  for (size_t i = 0; i < static_cast<size_t>(n); ++i)
+  for (int i = 0; i < n; ++i)
   {
-    for (size_t j = 0; j < static_cast<size_t>(m); ++j)
+    for (int j = 0; j < m; ++j)
     {
       assert(u0[j + m * i] == u1[j + m * i]);
     }
